Double type change to BigDecimal

+++ src/main/scala/com/dounine/tractor/behaviors/MarketTradeBehavior.scala
@@ -60,7 +60,7 @@ object MarketTradeBehavior extends ActorSerializerSuport {
       symbol: CoinSymbol,
       contractType: ContractType,
       direction: Direction,
-      price: Double,
+      price: BigDecimal,
       amount: Int,
       time: Long
   ) extends BaseSerializer
