+++ src/main/scala/com/dounine/tractor/behaviors/virtual/trigger/status/IdleStatus.scala
@@ -118,8 +118,8 @@ object IdleStatus extends JsonParse {
               offset: Offset,
               orderPriceType: OrderPriceType,
               triggerType: TriggerType,
-              orderPrice: Double,
-              triggerPrice: Double,
+              orderPrice: BigDecimal,
+              triggerPrice: BigDecimal,
               volume: Int
             ) => {
           logger.info(command.logJson)
